+++ src/lib.rs
@@ -50,6 +50,7 @@ pub mod stat;
 pub mod tags;
 pub mod task;
 pub mod tt;
+pub mod tt2;
 pub mod types;
 pub mod utils;
 pub mod version;